changed path from localhost to relative URL
the code i provided will not redirect to video.html on GitHub Pages because it was  referencing the http://127.0.0.1:5500 URL, which is a localhost address.

When we deploy our code to GitHub Pages, it runs in a different environment with a different domain. Therefore, we need to update the URL to point to the correct location on GitHub Pages.